# backend/docs/swagger.yaml
basePath: /api/v1
definitions:
...
  handlers.CreateOrderRequest:
    properties:
      cart_item_ids:
        description: 'Optional: specific cart items to order (as string UUIDs)'
        example:
        - 123e4567-e89b-12d3-a456-426614174000
        - 456e7890-e89b-12d3-a456-426614174001
        items:
          type: string
        type: array
      payment_method:
        enum:
        - credit_card
        - debit_card
        - paypal
        - bank_transfer
        example: credit_card
        type: string
      shipping_address:
        example: 123 Main St, City, State 12345
        maxLength: 500
        minLength: 10
        type: string
    required:
    - payment_method
    - shipping_address
    type: object
...
  handlers.UpdateOrderStatusRequest:
    properties:
      status:
        enum:
        - pending
        - processing
        - shipped
        - delivered
        - cancelled
        example: processing
        type: string
    required:
    - status
    type: object
...
  models.CartItem:
    properties:
      cart:
        allOf:
        - $ref: '#/definitions/models.ShoppingCart'
        description: Relationships
      cart_id:
        type: string
      created_at:
        type: string
      id:
        type: string
      product:
        $ref: '#/definitions/models.Product'
      product_id:
        type: string
      quantity:
        type: integer
      updated_at:
        type: string
    type: object
  models.Comment:
    properties:
      content:
        type: string
      created_at:
        type: string
      id:
        type: string
      product:
        $ref: '#/definitions/models.Product'
      product_id:
        type: string
      rating:
        description: 1-5 star rating
        type: integer
      updated_at:
        type: string
      user:
        allOf:
        - $ref: '#/definitions/models.User'
        description: Relationships
      user_id:
        type: string
    type: object
  models.Discount:
    properties:
      category:
        description: Nullable for product-specific discounts
        type: string
      created_at:
        type: string
      discount_type:
        description: '''percentage'', ''fixed_amount'''
        type: string
      discount_value:
        description: Percentage (0-100) or fixed amount
        type: number
      end_date:
        type: string
      id:
        type: string
      is_active:
        type: boolean
      max_discount_amount:
        description: Maximum discount amount (for percentage)
        type: number
      min_order_amount:
        description: Minimum order amount to apply discount
        type: number
      product:
        allOf:
        - $ref: '#/definitions/models.Product'
        description: Relationships
      product_id:
        description: Nullable for category-wide discounts
        type: string
      start_date:
        type: string
      updated_at:
        type: string
      usage_count:
        type: integer
      usage_limit:
        description: 0 = unlimited
        type: integer
    type: object
  models.Favorite:
    properties:
      created_at:
        type: string
      id:
        type: string
      product:
        $ref: '#/definitions/models.Product'
      product_id:
        type: string
      user:
        allOf:
        - $ref: '#/definitions/models.User'
        description: Relationships
      user_id:
        type: string
    type: object
  models.Order:
    properties:
      created_at:
        type: string
      id:
        type: string
      order_items:
        items:
          $ref: '#/definitions/models.OrderItem'
        type: array
      status:
        type: string
      total:
        type: number
      updated_at:
        type: string
      user:
        allOf:
        - $ref: '#/definitions/models.User'
        description: Relationships
      user_id:
        type: string
    type: object
  models.OrderItem:
    properties:
      created_at:
        type: string
      id:
        type: string
      order:
        allOf:
        - $ref: '#/definitions/models.Order'
        description: Relationships
      order_id:
        type: string
      price:
        type: number
      product:
        $ref: '#/definitions/models.Product'
      product_id:
        type: string
      quantity:
        type: integer
    type: object
  models.Product:
    properties:
      cart_items:
        items:
          $ref: '#/definitions/models.CartItem'
        type: array
      category:
        type: string
      comments:
        items:
          $ref: '#/definitions/models.Comment'
        type: array
      created_at:
        type: string
      description:
        type: string
      discounts:
        items:
          $ref: '#/definitions/models.Discount'
        type: array
      favorites:
        items:
          $ref: '#/definitions/models.Favorite'
        type: array
      id:
        type: string
      image_url:
        type: string
      name:
        type: string
      order_items:
        description: Relationships
        items:
          $ref: '#/definitions/models.OrderItem'
        type: array
      price:
        type: number
      product_views:
        items:
          $ref: '#/definitions/models.ProductView'
        type: array
      recommendations:
        items:
          $ref: '#/definitions/models.Recommendation'
        type: array
      stock:
        type: integer
      tags:
        items:
          $ref: '#/definitions/models.Tag'
        type: array
      updated_at:
        type: string
      upvotes:
        items:
          $ref: '#/definitions/models.Upvote'
        type: array
      user_interactions:
        items:
          $ref: '#/definitions/models.UserInteraction'
        type: array
    type: object
  models.ProductView:
    properties:
      created_at:
        description: Changed from Timestamp to CreatedAt for consistency
        type: string
      id:
        type: string
      product:
        $ref: '#/definitions/models.Product'
      product_id:
        type: string
      session_id:
        type: string
      user:
        allOf:
        - $ref: '#/definitions/models.User'
        description: Relationships
      user_id:
        type: string
    type: object
  models.Recommendation:
    properties:
      algorithm_type:
        description: '''collaborative'', ''content_based'', ''hybrid'''
        type: string
      created_at:
        type: string
      id:
        type: string
      product:
        $ref: '#/definitions/models.Product'
      product_id:
        type: string
      score:
        type: number
      user:
        allOf:
        - $ref: '#/definitions/models.User'
        description: Relationships
      user_id:
        type: string
    type: object
...
  models.ShoppingCart:
    properties:
      cart_items:
        items:
          $ref: '#/definitions/models.CartItem'
        type: array
      created_at:
        type: string
      id:
        type: string
      updated_at:
        type: string
      user:
        allOf:
        - $ref: '#/definitions/models.User'
        description: Relationships
      user_id:
        type: string
    type: object
  models.Tag:
    properties:
      color:
        description: Hex color code
        type: string
      created_at:
        type: string
      description:
        type: string
      id:
        type: string
      name:
        type: string
      products:
        description: Relationships
        items:
          $ref: '#/definitions/models.Product'
        type: array
    type: object
  models.Upvote:
    properties:
      created_at:
        type: string
      id:
        type: string
      product:
        $ref: '#/definitions/models.Product'
      product_id:
        type: string
      user:
        allOf:
        - $ref: '#/definitions/models.User'
        description: Relationships
      user_id:
        type: string
    type: object
  models.User:
    properties:
      comments:
        items:
          $ref: '#/definitions/models.Comment'
        type: array
      created_at:
        type: string
      email:
        type: string
      favorites:
        items:
          $ref: '#/definitions/models.Favorite'
        type: array
      id:
        type: string
      name:
        type: string
      orders:
        description: Relationships
        items:
          $ref: '#/definitions/models.Order'
        type: array
      recommendations:
        items:
          $ref: '#/definitions/models.Recommendation'
        type: array
      shopping_cart:
        $ref: '#/definitions/models.ShoppingCart'
      updated_at:
        type: string
      upvotes:
        items:
          $ref: '#/definitions/models.Upvote'
        type: array
      user_interactions:
        items:
          $ref: '#/definitions/models.UserInteraction'
        type: array
    type: object
  models.UserInteraction:
    properties:
      created_at:
        description: Changed from Timestamp to CreatedAt for consistency
        type: string
      id:
        type: string
      interaction_type:
        description: '''view'', ''cart_add'', ''purchase'', ''like'''
        type: string
      product:
        $ref: '#/definitions/models.Product'
      product_id:
        type: string
      session_id:
        type: string
      user:
        allOf:
        - $ref: '#/definitions/models.User'
        description: Relationships
      user_id:
        type: string
    type: object
host: localhost:8081
info:
  contact:
    email: support@bachelor-ecommerce.com
    name: API Support
    url: http://www.swagger.io/support
  description: ML-Powered E-commerce Platform API with advanced recommendation system,
    intelligent search, and comprehensive analytics
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: http://swagger.io/terms/
  title: Bachelor E-commerce API
  version: "1.0"
paths:
...
  /orders:
    get:
      consumes:
      - application/json
      description: Get paginated list of user's order history with order items
      parameters:
      - default: 1
        description: Page number
        in: query
        name: page
        type: integer
      - default: 10
        description: Items per page
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Orders retrieved successfully
          schema:
            additionalProperties: true
            type: object
        "401":
          description: User not authenticated
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Get user orders
      tags:
      - Orders
    post:
      consumes:
      - application/json
      description: Create a new order from the user's current cart items or specific
        cart items with atomic stock management
      parameters:
      - description: 'Order creation data. cart_item_ids is optional - if not provided,
          orders all cart items. Example: {\'
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handlers.CreateOrderRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Order created successfully
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Invalid request, empty cart, or insufficient stock
          schema:
            additionalProperties: true
            type: object
        "401":
          description: User not authenticated
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Cart not found
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Create order from cart
      tags:
      - Orders
  /orders/{id}:
    get:
      consumes:
      - application/json
      description: Get detailed information about a specific order
      parameters:
      - description: Order ID (UUID)
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Order retrieved successfully
          schema:
            $ref: '#/definitions/models.Order'
        "400":
          description: Invalid order ID
          schema:
            additionalProperties: true
            type: object
        "401":
          description: User not authenticated
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Order not found
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Get order by ID
      tags:
      - Orders
  /orders/{id}/cancel:
    put:
      consumes:
      - application/json
      description: Cancel a pending order and restore product stock
      parameters:
      - description: Order ID (UUID)
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Order cancelled successfully
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Invalid order ID or order cannot be cancelled
          schema:
            additionalProperties: true
            type: object
        "401":
          description: User not authenticated
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Order not found
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Cancel order
      tags:
      - Orders
  /orders/{id}/status:
    put:
      consumes:
      - application/json
      description: Update the status of an order with validation for status transitions
        (admin access required)
      parameters:
      - description: Order ID (UUID)
        in: path
        name: id
        required: true
        type: string
      - description: New order status
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handlers.UpdateOrderStatusRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Order status updated successfully
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Invalid request or status transition
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Order not found
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Update order status
      tags:
      - Orders
  /orders/stats:
    get:
      consumes:
      - application/json
      description: Get comprehensive order statistics for the authenticated user
      produces:
      - application/json
      responses:
        "200":
          description: Order statistics retrieved successfully
          schema:
            additionalProperties: true
            type: object
        "401":
          description: User not authenticated
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Get order statistics
      tags:
      - Orders
...
schemes:
- http
- https
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"
...